+++ src/__test__/Mail2Issue.handleCommentEvent.test.ts
@@ -5,7 +5,7 @@ import StateProvider from "../StateProvider";
 import { MessageTypes, Comment } from "../types";
 import issuesCommentsFixture from "./fixtures/comments.json";
 import fs from "fs";
-import FileStorageProvider from "../FileStorageProvider.ts1";
+import FileStorageProvider from "../FileStorageProvider.ts";
 
 const IssueFixture = {
   id: 10001,
